requirements.txt
setuptools
django
django-import-export
django-bootstrap-modal-forms
pillow
fpdf2
requests
